# api-contract/dist/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Fastlane E-Commerce API
  version: 0.5.0-draft
  summary: Contract-first REST API for the Fastlane B2C MVP
  description: |
    Week-1 API contract aligned with PRD Fastlane E-Commerce v1.0.0.
    This artifact defines the integration boundary before backend, storefront, admin, database, and infrastructure work are merged.
    All money values are non-negative whole-VND strings. Pricing, option compatibility, deposit schedules, promotion eligibility, and inventory are calculated on the server.
    Catalog content is available only in vi-VN. UI and error messages are localized by the client from stable error codes.
    Authentication and registration use Auth0 Universal Login; this API does not duplicate Auth0 login, registration, or logout endpoints.
  license:
    name: Internal training use only
    identifier: LicenseRef-Internal-Training
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /admin/products:
    get:
      tags:
        - Admin Catalog
      operationId: listAdminProducts
      x-prd-references:
        - US-A02
        - US-A08
        - FR-ADMN-02
        - FR-ADMN-03
      summary: List products for administration
      description: Search includes active and inactive products. sort=relevance requires q; when sort is omitted, q uses relevance then product ID and an empty q uses newest then product ID.
      security:
        - bearerAuth: []
      x-required-roles:
        - admin
      x-required-permissions:
        - catalog:manage
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/CategoryIdFilter'
        - $ref: '#/components/parameters/IsActive'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          $ref: '#/components/responses/ProductPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Admin Catalog
      operationId: createProduct
      x-prd-references:
        - US-A02
        - FR-ADMN-02
        - BR-05
        - BR-07
      summary: Create a complete accessory product aggregate
      description: |
        The currently implemented authoring workflow accepts ACCESSORY products. The backend validates the complete body before calling the database.
        Product, taxonomy memberships, service labels, option groups/values, SKU mappings and media are then written atomically.
        Prices belong to SKU variants; option values cannot add price. Existing inventory is never changed here and every new SKU starts at zero.
      security:
        - bearerAuth: []
      x-required-roles:
        - admin
      x-required-permissions:
        - catalog:manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccessoryWriteRequest'
      responses:
        '201':
          description: Accessory aggregate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessorySaveResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 RS256 JWT access token for this API. Deployment configuration supplies issuer, audience, JWKS URI, and namespaced roles claim.
        The backend validates signature, iss, aud, exp, role, permissions, and resource ownership. ID tokens are not accepted as API credentials.
  parameters:
    Query:
      name: q
      in: query
      description: Trimmed free-text query. Empty or whitespace-only values are invalid when supplied.
      schema:
        type: string
        minLength: 1
        maxLength: 120
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: relevance requires q. If omitted, q uses relevance and requests without q use newest.
      schema:
        type: string
        enum:
          - relevance
          - newest
          - price_asc
          - price_desc
...
    CategoryIdFilter:
      name: categoryId
      in: query
      schema:
        $ref: '#/components/schemas/Uuid'
    IsActive:
      name: isActive
      in: query
      schema:
        type: boolean
...
  responses:
    TooManyRequests:
      description: Rate limit exceeded.
      x-error-codes:
        - RATE_LIMIT_EXCEEDED
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Seconds before retrying.
        X-RateLimit-Limit:
          schema:
            type: integer
            minimum: 1
        X-RateLimit-Remaining:
          schema:
            type: integer
            minimum: 0
        X-RateLimit-Reset:
          schema:
            type: integer
            minimum: 0
          description: Unix epoch seconds when the window resets.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded.
              requestId: req_01J00000000000000000000000
    InternalServerError:
      description: Unexpected server error. Internal details are not exposed.
      x-error-codes:
        - INTERNAL_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: Unexpected server error.
              requestId: req_01J00000000000000000000000
...
    ProductPage:
      description: Product page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductPageResponse'
    ValidationError:
      description: Request syntax, parameters, or schema are invalid.
      x-error-codes:
        - VALIDATION_FAILED
        - DATE_RANGE_INVALID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_FAILED
              message: Request validation failed.
              requestId: req_01J00000000000000000000000
              fields:
                - path: shippingAddress.phoneNumber
                  code: INVALID_PHONE
...
    Unauthorized:
      description: Missing, invalid, or expired access token.
      x-error-codes:
        - AUTHENTICATION_REQUIRED
        - INVALID_ACCESS_TOKEN
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer authentication challenge.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
              requestId: req_01J00000000000000000000000
    Forbidden:
      description: Principal lacks the required role or at least one required permission.
      x-error-codes:
        - INSUFFICIENT_PERMISSION
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INSUFFICIENT_PERMISSION
              message: Permission is insufficient.
              requestId: req_01J00000000000000000000000
...
    BusinessRuleError:
      description: Request is well-formed but violates a business invariant.
      x-error-codes:
        - PRODUCT_NOT_PURCHASABLE
        - PRODUCT_CONFIGURATION_INVALID
        - ORDER_CANCELLATION_NOT_ALLOWED
        - PAYMENT_DEADLINE_EXPIRED
        - REFUND_NOT_ALLOWED
        - CART_EMPTY
        - BUSINESS_RULE_VIOLATION
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BUSINESS_RULE_VIOLATION
              message: Business rule was violated.
              requestId: req_01J00000000000000000000000
...
    Conflict:
      description: Request conflicts with current state or a unique constraint.
      x-error-codes:
        - CONFLICT
        - CATEGORY_IN_USE
        - DUPLICATE_CATEGORY_SLUG
        - DUPLICATE_PRODUCT_SLUG
        - DUPLICATE_SKU
        - DUPLICATE_PROMOTION_CODE
        - INVENTORY_VERSION_CONFLICT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: CONFLICT
              message: Request conflicts with current state.
              requestId: req_01J00000000000000000000000
    ServiceUnavailable:
      description: A required database capability has not been deployed or is temporarily unavailable.
      x-error-codes:
        - INTERNAL_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: Required catalog persistence capability is unavailable.
              requestId: req_01J00000000000000000000000
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - DATE_RANGE_INVALID
        - AUTHENTICATION_REQUIRED
        - INVALID_ACCESS_TOKEN
        - INSUFFICIENT_PERMISSION
        - RESOURCE_NOT_FOUND
        - CONFLICT
        - CATEGORY_IN_USE
        - DUPLICATE_CATEGORY_SLUG
        - DUPLICATE_PRODUCT_SLUG
        - DUPLICATE_SKU
        - DUPLICATE_PROMOTION_CODE
        - PRODUCT_NOT_PURCHASABLE
        - PRODUCT_CONFIGURATION_INVALID
        - ORDER_CANCELLATION_NOT_ALLOWED
        - PAYMENT_DEADLINE_EXPIRED
        - REFUND_NOT_ALLOWED
        - CART_EMPTY
        - CART_CHANGED
        - PRICE_CHANGED
        - OUT_OF_STOCK
        - PROMOTION_INVALID
        - PROMOTION_INACTIVE
        - PROMOTION_NOT_STARTED
        - PROMOTION_EXPIRED
        - PROMOTION_MINIMUM_NOT_MET
        - PROMOTION_USAGE_LIMIT_REACHED
        - IDEMPOTENCY_KEY_REUSED
        - INVENTORY_VERSION_CONFLICT
        - INVALID_ORDER_TRANSITION
        - TEST_DRIVE_SLOT_UNAVAILABLE
        - DUPLICATE_TEST_DRIVE_REQUEST
        - INVALID_TEST_DRIVE_TRANSITION
        - TEST_DRIVE_LOCATION_UNAVAILABLE
        - TEST_DRIVE_PRODUCT_UNAVAILABLE
        - INVALID_CONSULTATION_TRANSITION
        - ESTIMATE_POLICY_UNAVAILABLE
        - LOAN_PACKAGE_UNAVAILABLE
        - ESTIMATE_STALE
        - POLICY_VERSION_CONFLICT
        - SETTINGS_VERSION_CONFLICT
        - BUSINESS_RULE_VIOLATION
        - RATE_LIMIT_EXCEEDED
        - INTERNAL_ERROR
    ValidationFieldCode:
      type: string
      enum:
        - REQUIRED
        - UNKNOWN_FIELD
        - INVALID_FORMAT
        - INVALID_VALUE
        - INVALID_PHONE
        - OUT_OF_RANGE
        - TOO_SHORT
        - TOO_LONG
    ErrorField:
      type: object
      additionalProperties: false
      required:
        - path
        - code
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 200
          example: shippingAddress.phoneNumber
        code:
          $ref: '#/components/schemas/ValidationFieldCode'
        meta:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - requestId
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Safe presentation-neutral diagnostic; client localizes user-facing text from code.
            requestId:
              type: string
              minLength: 1
              maxLength: 100
            fields:
              type: array
              items:
                $ref: '#/components/schemas/ErrorField'
            meta:
              type: object
              additionalProperties: true
    Uuid:
      type: string
      format: uuid
    Category:
      type: object
      required:
        - id
        - name
        - slug
        - isActive
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          maxLength: 120
        slug:
          type: string
          maxLength: 160
        isActive:
          type: boolean
...
    Money:
      type: string
      pattern: ^(0|[1-9]\d*)$
      examples:
        - '199000'
      description: Non-negative whole-VND amount encoded as a base-10 string to avoid floating-point loss.
    ProductKind:
      type: string
      enum:
        - car
        - motorbike
        - accessory
      description: Stable business discriminator; category remains the mutable merchandising taxonomy.
    PurchaseMode:
      type: string
      enum:
        - full
        - deposit
    CancellationCutoff:
      type: string
      enum:
        - before_initial_payment
        - before_balance_due
        - before_shipping
    CancellationPolicy:
      type: object
      additionalProperties: false
      required:
        - customerCancellationAllowed
        - customerCancellationCutoff
        - refundPercentage
        - overdueRefundPercentage
        - cancellationFeeAmount
      properties:
        customerCancellationAllowed:
          type: boolean
        customerCancellationCutoff:
          allOf:
            - $ref: '#/components/schemas/CancellationCutoff'
          description: before_balance_due is valid only for deposit products; backend enforces the cross-field rule.
        refundPercentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Percentage of gross paid amount refundable for an allowed cancellation before overdue auto-cancellation.
        overdueRefundPercentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Percentage used when a deposit order is auto-cancelled after its grace period.
        cancellationFeeAmount:
          $ref: '#/components/schemas/Money'
    PositiveMoney:
      type: string
      pattern: ^[1-9]\d*$
      examples:
        - '199000'
      description: Strictly positive whole-VND amount.
    PurchaseTerms:
      description: Product-level payment, deadline, grace-period, cancellation, and refund policy. Order items snapshot this object at checkout.
      oneOf:
        - type: object
          additionalProperties: false
          required:
            - paymentMode
            - depositAmount
            - initialPaymentWindowMinutes
            - balancePaymentWindowDays
            - gracePeriodHours
            - cancellationPolicy
          properties:
            paymentMode:
              allOf:
                - $ref: '#/components/schemas/PurchaseMode'
              const: full
            depositAmount:
              type: 'null'
            initialPaymentWindowMinutes:
              type: integer
              minimum: 5
              maximum: 1440
            balancePaymentWindowDays:
              type: 'null'
            gracePeriodHours:
              type: 'null'
            cancellationPolicy:
              $ref: '#/components/schemas/CancellationPolicy'
        - type: object
          additionalProperties: false
          required:
            - paymentMode
            - depositAmount
            - initialPaymentWindowMinutes
            - balancePaymentWindowDays
            - gracePeriodHours
            - cancellationPolicy
          properties:
            paymentMode:
              allOf:
                - $ref: '#/components/schemas/PurchaseMode'
              const: deposit
            depositAmount:
              $ref: '#/components/schemas/PositiveMoney'
            initialPaymentWindowMinutes:
              type: integer
              minimum: 5
              maximum: 1440
            balancePaymentWindowDays:
              type: integer
              minimum: 1
              maximum: 365
              description: Measured from successful initial deposit payment.
            gracePeriodHours:
              type: integer
              minimum: 0
              maximum: 720
            cancellationPolicy:
              $ref: '#/components/schemas/CancellationPolicy'
    PriceRange:
      type: object
      additionalProperties: false
      description: Minimum and maximum base effective prices across active variants before dynamic option adjustments; active out-of-stock variants are included.
      required:
        - minimum
        - maximum
        - currency
      properties:
        minimum:
          $ref: '#/components/schemas/PositiveMoney'
        maximum:
          $ref: '#/components/schemas/PositiveMoney'
        currency:
          type: string
          enum:
            - VND
    CatalogCollectionKind:
      type: string
      enum:
        - CATEGORY
        - MODEL
        - CAMPAIGN
    CatalogVehicleFilterMode:
      type: string
      enum:
        - NONE
        - COLLECTION_MEMBERSHIP
        - VERIFIED_FITMENT
      description: COLLECTION_MEMBERSHIP is source taxonomy membership and is not a verified installation-fitment claim.
    CatalogVehicleModel:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - slug
        - name
        - vehicleKind
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          minLength: 1
          maxLength: 80
        slug:
          type: string
          minLength: 1
          maxLength: 160
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        name:
          type: string
          minLength: 1
          maxLength: 160
        vehicleKind:
          type: string
          enum:
            - CAR
            - MOTORBIKE
            - OTHER
    CatalogCollection:
      type: object
      additionalProperties: false
      required:
        - id
        - parentId
        - kind
        - sourceSystem
        - sourceKey
        - slug
        - name
        - vehicleFilterMode
        - displayOrder
        - vehicleModel
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        parentId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        kind:
          $ref: '#/components/schemas/CatalogCollectionKind'
        sourceSystem:
          type: string
          minLength: 1
          maxLength: 80
        sourceKey:
          type: string
          minLength: 1
          maxLength: 160
        slug:
          type: string
          minLength: 1
          maxLength: 160
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        name:
          type: string
          minLength: 1
          maxLength: 160
        vehicleFilterMode:
          $ref: '#/components/schemas/CatalogVehicleFilterMode'
        displayOrder:
          type: integer
          minimum: 0
        vehicleModel:
          oneOf:
            - $ref: '#/components/schemas/CatalogVehicleModel'
            - type: 'null'
    CatalogCollectionMembership:
      type: object
      additionalProperties: false
      required:
        - id
        - sourceSystem
        - isPrimary
        - firstSeenAt
        - lastSeenAt
        - collection
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        sourceSystem:
          type: string
          minLength: 1
          maxLength: 80
        isPrimary:
          type: boolean
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        collection:
          $ref: '#/components/schemas/CatalogCollection'
    ProductSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - slug
        - productKind
        - purchaseTerms
        - category
        - thumbnailUrl
        - priceRange
        - isActive
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 220
        productKind:
          $ref: '#/components/schemas/ProductKind'
        purchaseTerms:
          $ref: '#/components/schemas/PurchaseTerms'
        category:
          $ref: '#/components/schemas/Category'
        thumbnailUrl:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
          description: Non-null for public active products; may be null in admin draft results.
        priceRange:
          $ref: '#/components/schemas/PriceRange'
        isActive:
          type: boolean
        collectionMemberships:
          type: array
          items:
            $ref: '#/components/schemas/CatalogCollectionMembership'
          description: Optional normalized taxonomy projection during the v1 compatibility rollout.
    PageMeta:
      type: object
      additionalProperties: false
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ProductPageResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    AccessoryContentSectionType:
      type: string
      enum:
        - TECHNICAL_SPECS
        - FEATURES
        - USAGE_GUIDE
        - CARE_GUIDE
        - INSTALLATION_GUIDE
        - PACKAGE_CONTENTS
        - WARRANTY
        - SHIPPING_NOTE
        - SAFETY_NOTE
        - PURCHASE_NOTE
        - OTHER
    AccessoryContentAttribute:
      type: object
      additionalProperties: false
      required:
        - label
        - value
      properties:
        label:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
...
    AdminAccessoryContentSectionInput:
      type: object
      additionalProperties: false
      required:
        - key
        - type
        - title
        - displayOrder
        - body
        - items
        - attributes
      anyOf:
        - properties:
            body:
              type: string
              minLength: 1
        - properties:
            items:
              type: array
              minItems: 1
        - properties:
            attributes:
              type: array
              minItems: 1
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 80
          pattern: ^[a-z0-9_]+$
        type:
          $ref: '#/components/schemas/AccessoryContentSectionType'
        title:
          type: string
          minLength: 1
          maxLength: 200
        displayOrder:
          type: integer
          minimum: 10
          maximum: 300
          multipleOf: 10
        body:
          type:
            - string
            - 'null'
          minLength: 1
          maxLength: 20000
        items:
          type: array
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 2000
        attributes:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AccessoryContentAttribute'
    AdminAccessoryContentInput:
      type: object
      additionalProperties: false
      required:
        - schema
        - sections
      properties:
        schema:
          type: string
          const: accessory_content_v1
        sections:
          type: array
          maxItems: 30
          items:
            $ref: '#/components/schemas/AdminAccessoryContentSectionInput'
    AdminAccessoryOptionValueInput:
      type: object
      additionalProperties: false
      description: An option value identifies a SKU selection. It never adds price; price is stored on the resulting SKU variant.
      required:
        - code
        - name
        - colorHex
        - swatchUrl
        - displayOrder
        - imageUrls
      properties:
        existingId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          minLength: 1
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        name:
          type: string
          minLength: 1
          maxLength: 160
        colorHex:
          type:
            - string
            - 'null'
          pattern: ^#[0-9A-Fa-f]{6}$
        swatchUrl:
          type:
            - string
            - 'null'
          format: uri
          maxLength: 2048
        displayOrder:
          type: integer
          minimum: 0
        imageUrls:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
            maxLength: 2048
    AdminAccessoryOptionGroupInput:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - displayType
        - minimumSelections
        - maximumSelections
        - displayOrder
        - drivesMedia
        - values
      properties:
        existingId:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          minLength: 1
          maxLength: 80
          pattern: ^[a-z][a-z0-9_]*$
        name:
          type: string
          minLength: 1
          maxLength: 160
        displayType:
          type: string
          enum:
            - BUTTON
            - SWATCH
            - SELECT
        minimumSelections:
          type: integer
          enum:
            - 0
            - 1
        maximumSelections:
          type: integer
          const: 1
        displayOrder:
          type: integer
          minimum: 0
        drivesMedia:
          type: boolean
          description: At most one group may drive option-scoped media.
        values:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/AdminAccessoryOptionValueInput'
    AdminAccessoryVariantInput:
      type: object
      additionalProperties: false
      description: The sellable and inventory-owned SKU. originalPrice and salePrice are whole-VND JSON integers.
      required:
        - name
        - sku
        - originalPrice
        - salePrice
        - isActive
        - optionValues
        - imageUrls
      properties:
        existingId:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          minLength: 1
          maxLength: 200
        sku:
          type: string
          minLength: 1
          maxLength: 80
        originalPrice:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        salePrice:
          type:
            - integer
            - 'null'
          minimum: 0
          maximum: 9007199254740991
          description: Must be lower than originalPrice when present.
        isActive:
          type: boolean
        optionValues:
          type: object
          additionalProperties:
            type: string
            minLength: 1
            maxLength: 80
            pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          description: Exact group-code to value-code selections for this SKU. Every required group must be represented.
        imageUrls:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
            maxLength: 2048
    AdminAccessoryWriteRequest:
      type: object
      additionalProperties: false
      description: |
        Complete ACCESSORY aggregate replacement. The backend validates the entire body before opening the database write and then persists product, taxonomy, service labels, options, SKUs, mappings, inventory initialization and media in one transaction.
        Option values never add price. Existing variant IDs preserve inventory; new variants start at zero inventory. PATCH requires expectedUpdatedAt.
      required:
        - categoryId
        - primaryCollectionId
        - modelCollectionIds
        - name
        - slug
        - description
        - isActive
        - serviceLabelIds
        - content
        - optionGroups
        - variants
        - productImageUrls
      properties:
        expectedUpdatedAt:
          type: string
          format: date-time
          description: Required by PATCH for optimistic concurrency and omitted by POST.
        categoryId:
          $ref: '#/components/schemas/Uuid'
        primaryCollectionId:
          $ref: '#/components/schemas/Uuid'
        modelCollectionIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Uuid'
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 220
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        description:
          type: string
          minLength: 1
          maxLength: 10000
        isActive:
          type: boolean
        serviceLabelIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Uuid'
        content:
          $ref: '#/components/schemas/AdminAccessoryContentInput'
        optionGroups:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/AdminAccessoryOptionGroupInput'
        variants:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/AdminAccessoryVariantInput'
        productImageUrls:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
            maxLength: 2048
    AdminAccessorySaveResult:
      type: object
      additionalProperties: false
      required:
        - id
        - productType
        - isActive
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        productType:
          type: string
          const: ACCESSORY
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    AdminAccessorySaveResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/AdminAccessorySaveResult'
...